Changes Made to DB and addRecord
There was a sever issue with the records table in the DB, the
auto-increment would not allow an ID or SRID to be generated and two
auto-increments are not allowed on one table. A work around for this is
to have a random number generator on the addRecord form to created a 6
digit SRID. The table has been recreated from a previous version that
was on GitHub.