Resolved issue with inputService
It was previously working however due to a possibly corrupt DB issue
earlier and restored from a previous DB(?) it would keep prompting that
a user has no services left when they had 3 left. It appears to be the
now it wasn't a DBNull and was now a null value being returned from the
query. So far, it seems to be working.