Major work to inputService
Design changes to allow it to only display the SRID text box then pull
the values from the DB into some read only textboxes so the operator can
at least see the SRID is for the right person. There was a fatal error
that often happened if the if the services left values was null even
though the DB was meant to set it to true, to over come this an IF
statement is now in place so it the value is null then set it to true.